requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
cvzone==1.6.1
firebase-admin==6.5.0
python-dotenv==1.0.1
pillow==10.2.0
face-recognition==1.3.0